Import jquery with browserify
