+++ index.html
@@ -56,7 +56,6 @@
 
     <audio id=notice src="sound/nc32318.wav" preload=auto></audio>
 
-    <script src="bower_components/jquery/dist/jquery.min.js"></script>
     <script src="bower_components/socket.io-client/socket.io.js"></script>
     <script src="bower_components/angular/angular.js"></script>
     <script src="bower_components/angular-route/angular-route.js"></script>
@@ -68,11 +67,6 @@
     <script src="bower_components/angular-material/angular-material.js"></script>
     <script src="bower_components/angular-material-icons/angular-material-icons.js"></script>
     <script src="bower_components/dice3d/dist/dice3d.js"></script>
-    <script src="js/jquery.cssanim.js"></script>
-    <script src="js/socket.js"></script>
-    <script src="js/chat-message.js"></script>
-    <script src="js/chat-form.js"></script>
-    <script src="js/chat-writing.js"></script>
-    <script src="js/main.js"></script>
+    <script src="js/browser.js"></script>
 </body>
 </html>
