+++ src/browser/index.js
@@ -1,0 +1,6 @@
+require('./jquery.cssanim.js');
+require('./socket.js');
+require('./chat-message.js');
+require('./chat-form.js');
+require('./chat-writing.js');
+require('./main.js');
